added early return to allTasks selector. renamed to index
